requirements.txt
# Backend
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
python-dotenv==1.0.1
httpx==0.27.0

# AI / LLM
google-generativeai==0.5.4

# Frontend
streamlit==1.35.0
requests==2.32.2

# Utilities
python-jose==3.3.0

# Testing
pytest==8.2.0